Changed name of beacon sensor node
