+++ scripts/sensors/BeaconSensor.py
@@ -1,146 +1,142 @@
 #!/usr/bin/env python
 
-# @author: Gabriel Ciolac
+# @author: Patrick Gavigan
 # @author: Simon Yacoub
 
 # SUBSCRIBER:   none
 # PUBLISHER:    String object to 'beacons' node
 
 import rospy
-import time
 from bluepy.btle import Scanner
 from std_msgs.msg import String
 from reader import BeaconReader
-
-# This script is intended to transmit raw beacon distances to the perception translator
-
-averages = dict()
-
-'''
-    Read RSSIs using bluepy library,
-    returns RSSIs in a dictionary with MAC addresses as the key
-'''
-def __read_RSSI():
-    reader = BeaconReader()
-    ble_list = Scanner().scan(0.2)
-    found_beacons = dict()
-    for dev in ble_list:
-        if dev.addr in reader.read_beacons().keys():
-            found_beacons[dev.addr] = dev.rssi
-    return found_beacons
+import statistics
 
 
+# Polls for bluetooth devices, returns signal strength in a dict, indexed by
+# MAC address.
+# beaconParameters: Parameters of the beacons we care about
+# timeout: How long to wait when polling for beacons
+def pollBeacons(beaconParameters, timeout):
+    beaconScan = Scanner().scan(timeout)
+    # beaconScan = Scanner().scan(1)
+    foundBeacons = dict()
+    for beacon in beaconScan:
+        if beacon.addr in beaconParameters.keys():
+            foundBeacons[beacon.addr] = beacon.rssi
+    return foundBeacons
 
-'''
-    Calculates distances from given RSSIs
-    returns dictionary of distances with MAC addresses as the keys
-'''
-def __get_distance():
-    
-    reader = BeaconReader()
-    read_rssi = __read_RSSI()
-    MAC_ADDRs = read_rssi.keys()
-    if len(MAC_ADDRs) == 0:
-        return None
+# Convert the signal strength data to a distance (hopefully in meters)
+def calculateDistance(beaconData,beaconParameters):
+    macAddresses = beaconData.keys()
+    #if len(macAddresses) == 0:
+    #    return None
 
     distances = dict()
-    for MAC in MAC_ADDRs:
-        distances[MAC] = pow(10,(reader.read_beacons()[MAC][1] - read_rssi[MAC])/(10*reader.read_beacons()[MAC][0]))
-    
+    for mac in macAddresses:
+        environmentVariable = beaconParameters[mac][0]
+        measuredValue = beaconParameters[mac][1]
+        distances[mac] = pow(10,(measuredValue - beaconData[mac])/(10*environmentVariable))
     return distances
-
-def __calc_average(mac, distance):
-    global averages
     
-    averages[mac].insert(0, distance)
-    averages[mac].pop(5)
-    avg = sum(averages[mac])/5
-    if distance < avg * 1.25 and distance > avg * 0.75:
-        return True
-    return False
+def manageBeaconQueue(beaconHistory,newData,queueLength=10):
+    beaconNames = beaconHistory.keys()
+    newBeaconDataNames = newData.keys()
+
+    for currentBeacon in beaconNames:
+        currentQueue = beaconHistory[currentBeacon]
+
+        if currentBeacon in newBeaconDataNames:
+            currentQueue.append(newData[currentBeacon])
+        
+        currentQueue = removeOutliers(currentQueue)
+        
+        # If the queue is too long, pop the oldest element
+        if len(currentQueue) > queueLength:
+            currentQueue.pop(0)
+        
+    return beaconHistory
+
+    
+# Remove outliers from range list
+def removeOutliers(rangeList,queueLength=10):
+    if len(rangeList) > queueLength:
+        dataMean = statistics.mean(rangeList)
+        dataStd = statistics.stdev(rangeList)
+        maxValue = dataMean + (3 * dataStd)
+        minValue = dataMean - (3 * dataStd)
+    
+        for i in range(len(rangeList)):
+            if (rangeList[i] < minValue) or (rangeList[i] > maxValue):
+                rangeList[i] = -1       # Flag the outlier
+            
+    rangeList = filter(lambda a: a != -1, rangeList) 
+    return rangeList
+
+    
+# Get averages for the distances
+def calculateAverages(beaconData):
+    beaconNames = beaconData.keys()
+    beaconAverages = dict()
+    for beaconName in beaconNames:
+        if len(beaconData[beaconName]) > 1:
+            beaconAverages[beaconName] = statistics.mean(beaconData[beaconName])
+        elif len(beaconData[beaconName]) == 1:
+            beaconAverages[beaconName] = beaconData[beaconName][0]
+    return beaconAverages
+
+
+def publishData(pub, data):
+    message = ""
+    beaconNames = data.keys()
+    for beaconName in beaconNames:
+        message = message + "beacon(" + str(beaconName) + "," + str(data[beaconName]) + ")"
+    
+    rospy.loginfo(message)
+    pub.publish(message)
+        
+
+def runBeacons(pub,rate,beaconParameters,beaconHistory):
+    period = 1.0/rate
+    beaconData = pollBeacons(beaconParameters, period)
+    distances = calculateDistance(beaconData,beaconParameters)
+
+    beaconHistory = manageBeaconQueue(beaconHistory,distances)
+
+    # Calculate the average for each queue
+    beaconAverage = calculateAverages(beaconHistory)
+    
+    # Publish averages
+    publishData(pub,beaconAverage)
+    
 
 def rosMain():
-    global averages
     
+    # Set up the ROS node
     pub = rospy.Publisher('/sensors/Beacon', String, queue_size=5)
     rospy.init_node('bluetoothBeacons', anonymous=True)
-    rate = rospy.Rate(60)
+    frequency = 5
+    rate = rospy.Rate(frequency)
     
-    reader = BeaconReader()
-    beacons = reader.read_beacons()
+    # Get the beacon parameters from the file
+    beaconReader = BeaconReader()
+    beaconParameters = beaconReader.read_beacons()
     
+    # Dict for the beacon history data queues
+    beaconHistory = dict()
+    beaconNames = beaconParameters.keys()
+    for beacon in beaconNames:
+        beaconHistory[beacon] = list()
+    
+    # Run the node
     while not rospy.is_shutdown():
-        timer = time.time() # take time at start of loop
-        # Initialize values
-        samples = dict()
-        sumDist = dict()
-        prevdist = dict()
-        
-        # Set default values for dicts
-        for mac in beacons.keys():
-            sumDist[mac] = 0
-            samples[mac] = 10
-        
-        # Take 10 samples, get sum
-        count = 0
-        valid = True
-        while (count < 10):
-            distances = __get_distance()
-            
-            if not distances == None:
-                outlier = False
-                # Check all distances of macs, see if an obvious outlier
-                for mac in distances.keys():
-                    try:
-                        if abs(distances[mac] - prevdist[mac]) > 1:
-                            samples[mac] -= 1
-                            outlier = True
-                            rospy.loginfo("outlier detected data trashed")
-                    except KeyError:
-                        pass
-                    except TypeError:
-                        pass
-                
-                if not outlier:
-                    # Get sum for every different beacon
-                    for mac in distances.keys():
-                        prevdist[mac] = distances[mac]
-                        sumDist[mac] = sumDist[mac] + distances[mac]
-                #rate.sleep() #Even though sampling takes 2.5sec this line forces 10 second wait for the sampels to be released
-                count += 1
-            else:
-                valid = False
-        
-        if valid:
-            # Find average of 10 samples (1 per second)
-            avg = dict()
-            for mac in sumDist:
-                if not samples[mac] == 0: 
-                    avg[mac] = sumDist[mac]/samples[mac]
-                else:
-                    avg[mac] = 100
-                if avg[mac] == 0:
-                    avg[mac] = 100
-            
-            # Construct message to publish
-            MAC_ADDRs = avg.keys()
-            to_send = ''
-            for MAC in MAC_ADDRs:
-                #the $ sign is used for splitting the different mac addresses
-                to_send = to_send + str(MAC)+','+str(avg[MAC])+'$' 
-            
-            # If at least 1 beacon wasnt an outlier, then publish
-            if not to_send == '':
-                timer = time.time() - timer # record the time between start of loop and publish perception
-                rospy.loginfo("Time taken to publish perception: " + str(timer))
-                rospy.loginfo(to_send)
-                pub.publish(to_send)
+        runBeacons(pub,frequency,beaconParameters,beaconHistory)
+        rate.sleep()
+    
 
 if __name__ == '__main__':
     try:
         rosMain()
     except rospy.ROSInterruptException:
         pass
-
-
